refactor: separate tg related logic
Separate telegram related logic to special folder for more modular
project layout
