+++ bot/tg/handlers/commands/help.py
@@ -2,8 +2,8 @@ from telegram import Update
 from telegram.constants import ChatAction
 from telegram.ext import ContextTypes
 
-from decorators.command_handler import command_handler
-from utils.constants import HELP_KEYBOARD_BUTTON
+from tg.decorators.command_handler import command_handler
+from tg.utils.constants import HELP_KEYBOARD_BUTTON
 
 
 with open("resources/demo.mp4", "rb") as file:
