+++ bot/tg/utils/constants.py
